refactor: Risk LCD useGameEngine(1,033行) をフェーズ別フックに分割
- useRunningPhase: ゲームサイクル・被弾判定・アナウンス・移動
- usePerkPhase: パーク選択画面の表示・選択実行
- useShopPhase: スタイル/ショップ/ヘルプ画面の入力処理
- useResultPhase: ゲーム終了・リザルト・タイトル復帰
- useGameEngine を各フェーズフックを統合するオーケストレーターに整理
- 循環依存は useRef パターンで解消
- 全53テスト通過・ビルド確認済み
